refactor(mcp): move test_scope out of approval_test_support (#5970 review)
test_scope is a generic ResourceScope fixture used by factory.rs and
runtime.rs's tests, but it lived in the approval-flavored
approval_test_support module (whose own comment already flagged the
mismatch). Split it into a dedicated scope_test_support module,
mirroring the crate's #[cfg(test)] mod pattern, and repoint both
call sites.
